fix(LabelMap): Sort labels numerically rather than alphabetically
Labels are assumed to be sorted numerically when the first and last
elements are used as the min and max, respectively, in a call to
setMappingRange on the color transfer function.  This fixes the
issue we have seen where picking indicated that several label
values were assigned the same color.