fix(viewer): Various rendering/ui improvements
Detect when we have only a labelmap and make adjustments to how user-
specified weights (from the labelmap piecewise function slider) are
mapped to the actual pwf.  Also, adjust weight->pwf mapping when we
are showing slice compared to volume, as we always want some context
in slice mode (so we can see segments to adjust pwf for).

Decouple the component selection checkboxes from the labelmap weight
slider, as those do not interact well together.  After this change,
the component visibility state consists of a boolen indicating whether
the component is shown or hidden, as well as float specifying the
weight.