feat(imJoy): Enable passing uiMachineOption's as an ES module
Many updates here.

The uiMachineOptions can be passed as an ES module, including through
the ImJoy plugin.

A custom UI can be built with Rollup as with the test.

The reference UI is now available as its own package,
itk-viewer-reference-ui. New UI's packages should follow the naming convention:

  itk-viewer-NAME-ui

where NAME is the name of the UI.