ci(Data): Use tarball for CI fetch
The IPFS node has difficulty downloading the data without a timeout.
This may be due to starting up the node and immediately requesting large
amounts of data without the opportunity to provide data to peers.

Download and extract a versioned tarball for now from GitHub Releases.

In the future, we will look to download a versioned, provenance rich
DataLad dataset.